requirements.txt
fastapi
uvicorn
openai
pydantic
google-api-python-client
openai-whisper
python-dotenv
